Namestena pretraga gradova kao i njihov odabir

+++ WeatherForecastApp/WeatherForecastApp/Main.cs
@@ -9,7 +9,7 @@ namespace WeatherForecastApp
     public class Main
     {
         public double temp { get; set; }
-        public int pressure { get; set; }
+        public double pressure { get; set; }
         public int humidity { get; set; }
         public double temp_min { get; set; }
         public double temp_max { get; set; }
